Rework landing example grid
