+++ apps/typegpu-docs/astro.config.mjs
@@ -40,10 +40,6 @@ export default defineConfig({
   markdown: {
     remarkPlugins: [remarkMath],
     rehypePlugins: [rehypeMathJax],
-    shikiConfig: {
-      theme: 'one-dark-pro',
-      wrap: true,
-    },
   },
   vite: {
     define: {
