--- apps/typegpu-docs/src/pages/landing-examples/3d-fish/index.html
@@ -1,53 +1,0 @@
-<style>
-  .controls-header {
-    font-size: 1.1rem;
-    margin-bottom: 0.5rem;
-  }
-
-  .controls {
-    font-size: 0.8rem;
-  }
-
-  #attribution {
-    margin-top: 1rem;
-    font-size: 0.8rem;
-  }
-
-  #help {
-    opacity: 0;
-    position: absolute;
-    margin: 0 auto;
-    text-align: left;
-    z-index: 1;
-    background-color: rgba(0, 0, 0, 0.7);
-    color: white;
-    padding: 0.625rem;
-    border-radius: 0.625rem;
-    user-select: none;
-    pointer-events: none;
-    transition: opacity 0.5s;
-  }
-
-  .spinner-background {
-    inset: 0;
-    position: absolute;
-    display: grid;
-    place-items: center;
-    background-color: var(--color-tameplum-50, white);
-    z-index: 4;
-  }
-</style>
-
-<div class="spinner-background">
-  <div class="spinner">Loading...</div>
-</div>
-<div id="help">
-  <h3 class="controls-header">Controls (click to dismiss)</h3>
-  <p class="controls"><b>Left Mouse Button:</b> Look around</p>
-  <p class="controls"><b>Right Mouse Button:</b> Repel fish from the cursor</p>
-  <p id="attribution">
-    Uses "Animated Low Poly Fish" by Sketchfab user Atlas, used under CC BY 4.0 / Modified from
-    original
-  </p>
-</div>
-<canvas data-fit-to-container></canvas>
