--- apps/typegpu-docs/src/pages/landing-examples/3d-fish/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "3D Fish",
-  "category": "rendering",
-  "tags": ["experimental", "3d", "rasterization", "compute", "double buffering"]
-}
