--- apps/typegpu-docs/src/pages/landing-examples/caustics/index.html
@@ -1,1 +1,0 @@
-<canvas></canvas>
