--- apps/typegpu-docs/src/pages/landing-examples/caustics/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Caustics",
-  "category": "rendering",
-  "tags": ["experimental", "fragment shader"]
-}
