--- apps/typegpu-docs/src/pages/landing-examples/cubemap-reflection/index.html
@@ -1,41 +1,0 @@
-<style>
-  .controls-header {
-    font-size: 1.1rem;
-    margin-bottom: 0.5rem;
-  }
-
-  .controls {
-    font-size: 0.8rem;
-  }
-
-  #attribution {
-    margin-top: 1rem;
-    font-size: 0.8rem;
-  }
-
-  #help {
-    opacity: 1;
-    position: absolute;
-    margin: 0 auto;
-    text-align: left;
-    z-index: 1;
-    background-color: rgba(0, 0, 0, 0.7);
-    color: white;
-    padding: 10px;
-    border-radius: 10px;
-    user-select: none;
-    pointer-events: none;
-    transition: opacity 0.5s;
-  }
-</style>
-<div id="help">
-  <h3 class="controls-header">Controls (click to dismiss)</h3>
-  <p class="controls"><b>Left Mouse Button:</b> Look around</p>
-  <p class="controls"><b>Scroll Wheel:</b> Zoom in/out</p>
-
-  <p id="attribution">
-    Uses textures by Emil "Humus" Persson <br />
-    under CC BY 4.0 / Modified from original
-  </p>
-</div>
-<canvas data-fit-to-container></canvas>
