--- apps/typegpu-docs/src/pages/landing-examples/cubemap-reflection/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Cubemap Reflection",
-  "category": "rendering",
-  "tags": ["experimental", "3d", "rasterization", "compute"]
-}
