--- apps/typegpu-docs/src/pages/landing-examples/jelly-slider/index.html
@@ -1,36 +1,0 @@
-<style>
-  #attribution {
-    opacity: 1;
-    position: absolute;
-    bottom: 0.625rem;
-    left: 50%;
-    transform: translateX(-50%);
-    z-index: 2;
-    background-color: rgba(0, 0, 0, 0.7);
-    color: white;
-    padding: 0.75rem 1rem;
-    border-radius: 0.625rem;
-    user-select: none;
-    pointer-events: auto;
-    transition: opacity 0.5s;
-    font-size: 1rem;
-    text-align: center;
-    white-space: nowrap;
-  }
-
-  #attribution a {
-    color: #87ceeb;
-    text-decoration: none;
-    cursor: pointer;
-  }
-
-  #attribution a:hover {
-    text-decoration: underline;
-  }
-</style>
-
-<div id="attribution">
-  Inspired by work of
-  <a href="https://x.com/cerpow/status/1964953851603358112" target="_blank">Voicu Apostol</a>
-</div>
-<canvas></canvas>
