--- apps/typegpu-docs/src/pages/landing-examples/jelly-slider/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Jelly slider",
-  "category": "rendering",
-  "tags": ["experimental"]
-}
