--- apps/typegpu-docs/src/pages/landing-examples/vaporrave/index.html
@@ -1,1 +1,0 @@
-<canvas data-fit-to-container></canvas>
