--- apps/typegpu-docs/src/pages/landing-examples/vaporrave/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Vaporrave",
-  "category": "rendering",
-  "tags": ["experimental", "sdf", "ray marching"]
-}
